Practica_1/RepartidorPython/requirements.txt
Flask==2.1.0
pyjwt==2.0.0
datetime
uuid
Flask-SQLAlchemy
itsdangerous==2.0.1
